Move benchmark report projection inward
